decls: PS2 parameter types where (int) casts bridged them: EmCatchSubSet(pEm, pSub, pl_dir, mode, x, y, z, void (*ft)(cSubChar*)) in the PS2 order, cMot3::set seq, cObjLadder::setDown / cObjPillar::setBreak/setThrow / cObj16::setPlDmgMot / em39BlendMotSet seq pointers (Seq_pl_escape / Seq_pl_dm void*), Obj01SetEst u32 eff/est parameters, cObjMgr::construct(cObj*, ID), SceKill(u32 level); 14 (int) casts dropped, the sym_map/symbols rows re-mangled (bytes unchanged)

+++ src/em3c/em3c.cpp
@@ -2262,7 +2262,7 @@ void em3cSetParasite(cEm3c* em)
     if (w->pCore) {
         w->pCore->setMotData(m0, m1, m2, m3, m4, m5, m6, m7, m8, m9, m10);
         w->pCore->setMotData(m0, m1, m2, m3, m4, m5, m6, m7, m8, m9, m10);
-        w->pCore->setPlDmgMot(ARC(0x61), (int) ARC(0x62));
+        w->pCore->setPlDmgMot(ARC(0x61), ARC(0x62));
     }
     if (female) {
         pos.x = 0.0f;
